% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.window.R
\name{age.window}
\alias{age.window}
\title{Age cut of survival data}
\usage{
age.window(dat, window, surv = c("enter", "exit", "event"))
}
\arguments{
\item{dat}{Input data frame. Must contain survival data.}

\item{window}{Vector of length two: The age interval.}

\item{surv}{Vector of length three giving tha names of the central
variables (a Surv object) in the dataframe 'dat'.}
}
\value{
A dataframe of the same form as the input, but 'cut' as
   desired. Intervals exceeding \code{window[2]} are cut there and
   given event = 0 (censoring).
}
\description{
For a given age interval, each spell is cut to fit into the
given age interval.
}
\examples{
dat <- data.frame(enter = 0, exit = 5.731, event = 1, x = 2)
window <- c(2, 5.3)
dat.trim <- age.window(dat, window)
export
}
\seealso{
\code{\link{cal.window}}.
}
\author{
Göran Broström
}
