% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.window.R
\name{cal.window}
\alias{cal.window}
\title{Calendar time cut of survival data}
\usage{
cal.window(dat, window, surv = c("enter", "exit", "event", "birthdate"))
}
\arguments{
\item{dat}{Input data frame. Must contain survival data.}

\item{window}{Vector of length two: The time interval.}

\item{surv}{Vector of length four giving tha names of the central
variables (an "extended" Surv object) in the dataframe 'dat'.}
}
\value{
A dataframe of the same form as the input, but 'cut' as
   desired. Intervals exceeding \code{window[2]} are cut there and
   given event = 0 (censoring).
}
\description{
For a given time interval, each spell is cut to fit into the
given calendar time interval.
}
\examples{
dat <- data.frame(enter = 0, exit = 5.731, event = 1,
birthdate = 1962.505, x = 2)
window <- c(1963, 1965)
dat.trim <- cal.window(dat, window)
}
\seealso{
\code{\link{age.window}}.
}
\author{
Göran Broström
}
