% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmean.R
\name{gmean}
\alias{gmean}
\title{Calculates the mean of Gompertz-Gamma distribution.}
\usage{
gmean(shape, scale, frail = 0)
}
\arguments{
\item{shape}{The shape parameter}

\item{scale}{The scale parameter}

\item{frail}{The frailty parameter}
}
\description{
Calculates the mean of Gompertz-Gamma distribution.
}
