% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gompreg.R
\name{gompreg}
\alias{gompreg}
\title{Maximizes a likelihood for the Gompertz-gamma distribution}
\usage{
gompreg(enter, exit, event)
}
\arguments{
\item{enter}{Late entrance}

\item{exit}{Time of failure or censoring}

\item{event}{Indicator of failure}
}
\description{
Maximizes a likelihood for the Gompertz-gamma distribution
}

