% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Ggompertz.R
\name{hGgompertz}
\alias{hGgompertz}
\title{The hazard function of a Gompertz-gamma distribution}
\usage{
hGgompertz(x, shape = 1, scale = 1, frail = 0, log.p = FALSE)
}
\arguments{
\item{x}{A vector of positive numbers}

\item{shape}{The \code{shape} parameter}

\item{scale}{The \code{scale} parameter}

\item{frail}{The \code{frailty} parameter}

\item{log.p}{Logical, should the result be on a log scale?}
}
\value{
A numerical vector of the same length as \code{x}.
}
\description{
The hazard function \eqn{h(x)} is defined as
\deqn{
h(x) = shape * exp(x / scale) / (1 + frail * scale * shape * (exp(x / scale) - 1))
}
When \eqn{frail = 0}, this simplifies to
\deqn{h(x) = shape * exp(x / scale)}
which is the hazard function of the standard Gompertz distribution
}
\examples{
hGgompertz(c(1, 3, 5))
}

