% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchreg2.R
\name{pchreg2}
\alias{pchreg2}
\title{Piecewise constant proportional hazards regression}
\usage{
pchreg2(X, Y, cuts, strata, offset, init, control)
}
\arguments{
\item{X}{The design (covariate) matrix.}

\item{Y}{A survival object, the response.}

\item{cuts}{Specifies the points in time where the hazard function
jumps. If omitted, an exponential model is fitted.}

\item{strata}{A stratum variable.}

\item{offset}{Offset.}

\item{init}{Initial regression parameter values.}

\item{control}{Controls convergence and output.}
}
\value{
A list
}
\description{
This function is called by \code{\link{phreg}}, but it can
also be directly called by a user.
}
\author{
Göran Broström
}
\seealso{
\code{\link{phreg}}.
}

