% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.fit.R
\name{phreg.fit}
\alias{phreg.fit}
\title{Parametric proportional hazards regression}
\usage{
phreg.fit(X, Y, dist, strata, offset, init, shape, control)
}
\arguments{
\item{X}{The design (covariate) matrix.}

\item{Y}{A survival object, the response.}

\item{dist}{Which baseline distribution?}

\item{strata}{A stratum variable.}

\item{offset}{Offset.}

\item{init}{Initial regression parameter values.}

\item{shape}{If positive, a fixed value of the shape parameter in the
distribution. Otherwise, the shape is estimated.If positive,
a fixed value of the shape parameter in the
distribution. Otherwise, the shape is estimated.}

\item{control}{Controls convergence and output.}
}
\value{
A list
}
\description{
This function is called by \code{\link{phreg}}, but it can also be directly called by
a user.
}
\author{
Göran Broström
}
\seealso{
\code{\link{phreg}}.
}

