% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{phreg}
\alias{phreg}
\title{Parametric Proportional Hazards Regression}
\usage{
phreg(formula = formula(data), data = parent.frame(),
  na.action = getOption("na.action"), dist = "weibull", cuts = NULL, init,
  shape = 0, control = list(eps = 1e-08, maxiter = 20, trace = FALSE),
  singular.ok = TRUE, model = FALSE, x = FALSE, y = TRUE)
}
\arguments{
\item{formula}{a formula object, with the response on the left of
a ~ operator, and the terms on the right. The response must be a survival
object as returned by the \code{Surv} function.}

\item{data}{a data.frame in which to interpret the variables named in
the formula.}

\item{na.action}{a missing-data filter function, applied to the model.frame,
after any subset argument has been used. Default is options()$na.action.}

\item{dist}{Which distribution? Default is "weibull", with the alternatives
"ev" (Extreme value), and "pch" (piecewise constant hazards function).
A special case like the exponential can be obtained by choosing "weibull"
in combination with shape = 1, or "pch" without cuts.}

\item{cuts}{Only used with dist = "pch". Specifies the points in time where
the hazard function jumps. If omitted, an exponential model is fitted.}

\item{init}{vector of initial values of the iteration. Default initial
value is zero for all variables.}

\item{shape}{If positive, the shape parameter is fixed at that value
(in each stratum). If zero or negative, the shape parameter is estimated.
If more than one stratum is present in data, each stratum gets its own
estimate.}

\item{control}{a list with components eps (convergence criterion),
maxiter (maximum number of iterations), and silent (logical,
controlling amount of output). You can change any component without
mention the other(s).}

\item{singular.ok}{Not used.}

\item{model}{Not used.}

\item{x}{Return the design matrix in the model object?
Default is \code{FALSE}}

\item{y}{Return the response in the model object? Default is \code{TRUE}.}
}
\value{
A list of class "phreg"
}
\description{
Proportional hazards model with parametric baseline hazard(s). Allows for
stratification with different scale and shape in each stratum, and left
truncated and right censored data.
}
\note{
The \code{gompertz} distribution is not included here.
}
\examples{
fit <- phreg(Surv(enter, exti, event) ~ ses, data = mort)
fit
}
\author{
Göran Broström
}
