% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ggompertz.R
\name{SGgompertz}
\alias{SGgompertz}
\title{The survival function of a Gompertz-gamma distribution}
\usage{
SGgompertz(x, shape = 1, scale = 1, frail = 0, log.p = FALSE)
}
\arguments{
\item{x}{A vector of positive numbers}

\item{shape}{The \emph{shape} parameter}

\item{scale}{The \emph{scale} parameter}

\item{frail}{The \emph{frailty} parameter}

\item{log.p}{Logical, should the result be on a log scale?}
}
\value{
A numerical vector of the same length as \code{x}.
}
\description{
The survival function of a Gompertz-gamma distribution
}
\details{
The survivor function \eqn{S(x)} is defined as
}
\examples{
SGgompertz(c(1, 3, 5))
}
