% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchreg3.R
\name{pchreg3}
\alias{pchreg3}
\title{Piecewise constant proportional hazards regression with aggregate}
\usage{
pchreg3(X, Y, cuts, strata = rep(1, NROW(Y)), offset = rep(0, NROW(Y)),
  init = rep(0, NCOL(as.matrix(X))), control = list(eps = 1e-08, maxiter =
  15, trace = TRUE))
}
\arguments{
\item{X}{The design (covariate) matrix.}

\item{Y}{A survival object, the response.}

\item{cuts}{Specifies the points in time where the hazard function
jumps. If omitted, an exponential model is fitted.}

\item{strata}{A stratum variable.}

\item{offset}{Offset.}

\item{init}{Initial regression parameter values.}

\item{control}{Controls convergence and output.}
}
\value{
A list
}
\description{
This function is called by \code{\link{phreg}}, but it can
also be directly called by a user.
}
\seealso{
\code{\link{phreg}}.
}
\author{
Göran Broström
}
